requirements-cloud.txt
# Cloud-specific Python dependencies
# (extends requirements.txt for GCP deployments)

# GCP SDK
google-cloud-storage>=2.14.0
google-cloud-run>=0.10.0
google-cloud-secret-manager>=2.18.0

# Vertex AI (Phase 3)
google-cloud-aiplatform>=1.43.0

# HTTP for health checks
uvicorn>=0.27.0
fastapi>=0.109.0

# pandas-ta for enrichment (already used but not in requirements.txt)
pandas-ta>=0.3.14b1
